Moved theme switch to home folder
